Changes to the javascript functions. Separated the Data from the functions.
Begin to look at the object oriented nature of data and functions. Are the functions necessary or implied in the object data itself?

Next up: begin building the site.

For future consideration: creating an true JavaScript object through the use of the Object constructor to enapsulate the data into a proper object with getters and setters.